ADD integration test into own class from controller test class ADD edge case test into own class from controller test class
